src/sage/symbolic/integration/integral.py: delete a doctest
We have a doctest that is testing for an unevaluated integral:

  sage: A = integral(1/ ((x-4) * (x^4+x+1)), x); A
  integrate(1/((x^4 + x + 1)*(x - 4)), x)

This is now failing because we output the maxima answer,

  -1/261*integrate((x^3 + 4*x^2 + 16*x + 65)/(x^4 + x + 1), x)
  + 1/261*log(x - 4)

rather than the giac answer, which is completely unevaluated. In an
earlier commit, giac was removed from the default list of integration
backends whenever giac is not actually installed because this produces
better error messages. But, as a side effect, we are no longer
skipping over the partial maxima answer to accept the unevaluated giac
answer. (This is actually an improvement in the case where giac is not
installed and where sympy raises an error.)

There is no important result tested by these lines, so I think the
least confusing thing to do is re-word the surrounding paragraph a bit
and delete this test. There are still other tests nearby for partial
evaluation of hard integrals.